Personal testing

+++ src/components/Home.js
@@ -1,4 +1,4 @@
-import React, {useState, useEffect} from "react";
+import React, {useEffect} from "react";
 import './styles/Home.css';
 // import MainQue from "./MainQue";
 import Button from "./Button"
@@ -20,10 +20,12 @@ const startQueue = () => {
   window.location = WEB_URL + "/Queue";
 };
 
-
-
 function Home() {
-  const [token, setToken] = useState("");
+  const enterPressed = e => {
+  if (e.key === "Enter" && document.getElementById(inputID).value.length === HASH_LENGTH) {
+    joinQueue();
+  }
+};
   const getReturnedParamsFromSpotifyAuth = (hash) => {
     const stringAfterHashtag = hash.substring(1);
     const paramsInUrl = stringAfterHashtag.split("&");
@@ -37,12 +39,13 @@ function Home() {
   };
   useEffect(() => {
     const { access_token, expires_in, token_type } = getReturnedParamsFromSpotifyAuth(window.location.hash);
-    localStorage.clear();
-    localStorage.setItem("token", access_token);
-    setToken(localStorage.getItem("token"));
-    localStorage.setItem("expiresIn", expires_in);
-    localStorage.setItem("tokenType", token_type);
-    window.history.pushState({}, document.title, "/");
+    if(access_token !== undefined && expires_in !== undefined && token_type !== undefined){
+      localStorage.clear();
+      localStorage.setItem("token", access_token);
+      localStorage.setItem("expiresIn", expires_in);
+      localStorage.setItem("tokenType", token_type);
+      window.history.pushState({}, document.title, "/home");
+    }
   }, [])
   const urlParams = new URLSearchParams(window.location.search);
   if(urlParams.get("error")){
@@ -51,7 +54,7 @@ function Home() {
     return (
       <div >
       <h1 ><Button text = "Join Existing Queue" onClick = {joinQueue} /> </h1>
-      <input type="text" name="queueID" id= {inputID} placeholder="Enter Queue ID" minLength= {HASH_LENGTH} maxLength={HASH_LENGTH}/>
+      <input type="text" name="queueID" id= {inputID} placeholder="Enter Queue ID" minLength= {HASH_LENGTH} maxLength={HASH_LENGTH}  onKeyPress={enterPressed}/>
       <h2><Button text = "Start a Queue" onClick = {startQueue} /> </h2>
       </div>
     );
